scheduling a job to refersh model periodically
