+++ src/predict/requirements.txt
@@ -1,9 +1,11 @@
 absl-py==0.10.0
+APScheduler==3.6.3
 astunparse==1.6.3
 atomicwrites==1.4.0
 attrs==19.3.0
 cachetools==4.1.1
 certifi==2020.6.20
+cffi==1.14.3
 chardet==3.0.4
 click==7.1.2
 cloudpickle==1.6.0
@@ -15,6 +17,7 @@ Flask==1.1.1
 Flask-Caching==1.9.0
 Flask-Cors==3.0.8
 Flask-PyMongo==2.3.0
+fsspec==0.8.4
 gast==0.3.3
 gcsfs==0.6.2
 geographiclib==1.50
@@ -25,6 +28,7 @@ google-auth-oauthlib==0.4.1
 google-cloud-bigquery==1.26.1
 google-cloud-core==1.3.0
 google-cloud-storage==1.30.0
+google-crc32c==1.0.0
 google-pasta==0.2.0
 google-resumable-media==0.6.0
 googleapis-common-protos==1.52.0
@@ -57,6 +61,7 @@ protobuf==3.12.2
 py==1.9.0
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
+pycparser==2.20
 pymongo==3.10.1
 pyparsing==2.4.7
 pytest==4.6.11
@@ -68,7 +73,6 @@ requests-oauthlib==1.3.0
 rsa==4.6
 scikit-learn==0.21.3
 scipy==1.5.1
-setuptools==41.6.0
 six==1.15.0
 statsmodels==0.10.1
 tabulate==0.8.7
@@ -79,9 +83,9 @@ tensorflow-estimator==2.3.0
 tensorflow-probability==0.11.1
 termcolor==1.1.0
 typing-extensions==3.7.4.3
+tzlocal==2.1
 urllib3==1.25.9
 wcwidth==0.2.5
 Werkzeug==1.0.1
-wheel==0.32.3
 wrapt==1.12.1
 zipp==3.1.0
